+++ src/App.jsx
@@ -3,35 +3,16 @@ import Header from "./layout/Header";
 import Footer from "./layout/Footer";
 import PageContent from "./layout/PageContent";
 import { BrowserRouter as Router } from "react-router-dom";
-import { api } from "./pages/SignupPage";
 import { useDispatch, useSelector } from "react-redux";
-import { setUser } from "./store/actions/clientActions";
+import { fetchToken } from "./store/actions/thunkActions";
 
 function App() {
   const dispatch = useDispatch();
   const user = useSelector((state) => state.client.user);
 
   useEffect(() => {
-    const fetchToken = async () => {
-      try {
-        if (user?.name) return; // Kullanıcı zaten varsa API isteği yapma
-        const token = localStorage.getItem("token"); // Token'ı güvenli bir şekilde al
-
-        if (!token) {
-          return;
-        }
-
-        const response = await api.get("/verify", {
-          headers: { Authorization: token },
-        });
-
-        dispatch(setUser(response.data)); // Redux store'a kullanıcıyı kaydet
-      } catch (error) {
-        console.error("Error fetching user:", error);
-      }
-    };
-
-    fetchToken();
+    if (user?.name) return; // Kullanıcı zaten varsa API isteği yapma
+    dispatch(fetchToken);
   }, [dispatch, user]); // user ve dispatch bağımlılık olarak eklendi
 
   return (
